+++ src/AuthorizeFilter.php
@@ -1,19 +1,19 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2;
+namespace virginent\oauth2;
 
-use conquer\oauth2\responsetypes\Authorization;
+use virginent\oauth2\responsetypes\Authorization;
 use Yii;
 use yii\base\ActionFilter;
 
 /**
  *
- * @author Andrey Borodulin
+ * @Author Daniel Lucas
  *
  */
 class AuthorizeFilter extends ActionFilter
@@ -21,8 +21,8 @@ class AuthorizeFilter extends ActionFilter
     private $_responseType;
 
     public $responseTypes = [
-        'token' => 'conquer\oauth2\responsetypes\Implicit',
-        'code' => 'conquer\oauth2\responsetypes\Authorization',
+        'token' => 'virginent\oauth2\responsetypes\Implicit',
+        'code' => 'virginent\oauth2\responsetypes\Authorization',
     ];
 
     /**
@@ -52,12 +52,12 @@ class AuthorizeFilter extends ActionFilter
     public function beforeAction($action)
     {
         if (!$responseType = BaseModel::getRequestValue('response_type')) {
-            throw new Exception(Yii::t('conquer/oauth2', 'Invalid or missing response type.'));
+            throw new Exception(Yii::t('virginent/oauth2', 'Invalid or missing response type.'));
         }
         if (isset($this->responseTypes[$responseType])) {
             $this->_responseType = Yii::createObject($this->responseTypes[$responseType]);
         } else {
-            throw new Exception(Yii::t('conquer/oauth2', 'An unsupported response type was requested.'), Exception::UNSUPPORTED_RESPONSE_TYPE);
+            throw new Exception(Yii::t('virginent/oauth2', 'An unsupported response type was requested.'), Exception::UNSUPPORTED_RESPONSE_TYPE);
         }
 
         $this->_responseType->validate();
@@ -88,7 +88,7 @@ class AuthorizeFilter extends ActionFilter
 
     /**
      * @throws Exception
-     * @return \conquer\oauth2\BaseModel
+     * @return \virginent\oauth2\BaseModel
      */
     protected function getResponseType()
     {
@@ -96,7 +96,7 @@ class AuthorizeFilter extends ActionFilter
             if (Yii::$app->session->has($this->storeKey)) {
                 $this->_responseType = unserialize(Yii::$app->session->get($this->storeKey));
             } else {
-                throw new Exception(Yii::t('conquer/oauth2', 'Invalid server state or the User Session has expired.'), Exception::SERVER_ERROR);
+                throw new Exception(Yii::t('virginent/oauth2', 'Invalid server state or the User Session has expired.'), Exception::SERVER_ERROR);
             }
         }
         return $this->_responseType;
@@ -112,7 +112,7 @@ class AuthorizeFilter extends ActionFilter
         /** @var Authorization $responseType */
         $responseType = $this->getResponseType();
         if (Yii::$app->user->isGuest) {
-            $responseType->errorRedirect(Yii::t('conquer/oauth2', 'The User denied access to your application.'), Exception::ACCESS_DENIED);
+            $responseType->errorRedirect(Yii::t('virginent/oauth2', 'The User denied access to your application.'), Exception::ACCESS_DENIED);
         }
         $parts = $responseType->getResponseData();