+++ src/Module.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2;
+namespace virginent\oauth2;
 
-use conquer\oauth2\console\Oauth2Controller;
+use virginent\oauth2\console\Oauth2Controller;
 use yii\base\BootstrapInterface;
 
 /**
- * @author Andrey Borodulin
+ * @author Andrey virginent
  */
 class Module extends \yii\base\Module implements BootstrapInterface
 {