+++ src/OAuth2IdentityInterface.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2;
+namespace virginent\oauth2;
 
 use yii\web\IdentityInterface;
 
 /**
  * Interface OAtuh2IdentityInterface
- * @package conquer\oauth2
+ * @package virginent\oauth2
  * @author Dmitry Fedorenko
  */
 interface OAuth2IdentityInterface