+++ src/TokenAction.php
@@ -1,18 +1,18 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2;
+namespace virginent\oauth2;
 
 use Yii;
 use yii\base\Action;
 use yii\web\Response;
 
 /**
- * @author Andrey Borodulin
+ * @Author Daniel Lucas
  */
 class TokenAction extends Action
 {
@@ -22,11 +22,11 @@ class TokenAction extends Action
     public $format = Response::FORMAT_JSON;
 
     public $grantTypes = [
-        'authorization_code' => 'conquer\oauth2\granttypes\Authorization',
-        'refresh_token' => 'conquer\oauth2\granttypes\RefreshToken',
-        'client_credentials' => 'conquer\oauth2\granttypes\ClientCredentials',
-//         'password' => 'conquer\oauth2\granttypes\UserCredentials',
-//         'urn:ietf:params:oauth:grant-type:jwt-bearer' => 'conquer\oauth2\granttypes\JwtBearer',
+        'authorization_code' => 'virginent\oauth2\granttypes\Authorization',
+        'refresh_token' => 'virginent\oauth2\granttypes\RefreshToken',
+        'client_credentials' => 'virginent\oauth2\granttypes\ClientCredentials',
+//         'password' => 'virginent\oauth2\granttypes\UserCredentials',
+//         'urn:ietf:params:oauth:grant-type:jwt-bearer' => 'virginent\oauth2\granttypes\JwtBearer',
     ];
 
     public function init()
@@ -38,12 +38,12 @@ class TokenAction extends Action
     public function run()
     {
         if (!$grantType = BaseModel::getRequestValue('grant_type')) {
-            throw new Exception(Yii::t('conquer/oauth2', 'The grant type was not specified in the request.'));
+            throw new Exception(Yii::t('virginent/oauth2', 'The grant type was not specified in the request.'));
         }
         if (isset($this->grantTypes[$grantType])) {
             $grantModel = Yii::createObject($this->grantTypes[$grantType]);
         } else {
-            throw new Exception(Yii::t('conquer/oauth2', 'An unsupported grant type was requested.'), Exception::UNSUPPORTED_GRANT_TYPE);
+            throw new Exception(Yii::t('virginent/oauth2', 'An unsupported grant type was requested.'), Exception::UNSUPPORTED_GRANT_TYPE);
         }
 
         $grantModel->validate();