+++ src/granttypes/JwtBearer.php
@@ -1,19 +1,19 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2\granttypes;
+namespace virginent\oauth2\granttypes;
 
-use conquer\oauth2\BaseModel;
+use virginent\oauth2\BaseModel;
 
 /**
  *
  * @link http://tools.ietf.org/html/draft-ietf-oauth-jwt-bearer-04#section-4
  *
- * @author Andrey Borodulin
+ * @Author Daniel Lucas
  */
 class JwtBearer extends BaseModel
 {