+++ src/models/Client.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2\models;
+namespace virginent\oauth2\models;
 
 use Yii;
 use yii\behaviors\BlameableBehavior;
@@ -66,11 +66,11 @@ class Client extends \yii\db\ActiveRecord
     public function attributeLabels()
     {
         return [
-            'client_id' => Yii::t('conquer/oauth2', 'Unique client identifier'),
-            'client_secret' => Yii::t('conquer/oauth2', 'Client secret'),
-            'redirect_uri' => Yii::t('conquer/oauth2', 'Redirect URI used for Authorization Grant'),
-            'grant_type' => Yii::t('conquer/oauth2', 'Space-delimited list of grant types permitted, null = all'),
-            'scope' => Yii::t('conquer/oauth2', 'Space-delimited list of approved scopes'),
+            'client_id' => Yii::t('virginent/oauth2', 'Unique client identifier'),
+            'client_secret' => Yii::t('virginent/oauth2', 'Client secret'),
+            'redirect_uri' => Yii::t('virginent/oauth2', 'Redirect URI used for Authorization Grant'),
+            'grant_type' => Yii::t('virginent/oauth2', 'Space-delimited list of grant types permitted, null = all'),
+            'scope' => Yii::t('virginent/oauth2', 'Space-delimited list of approved scopes'),
         ];
     }