+++ src/responsetypes/Authorization.php
@@ -1,18 +1,18 @@
 <?php
 /**
- * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
- * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
+ * @link https://github.com/virginent/yii2-oauth2-server
+ * @copyright Copyright (c) 2021 Daniel Lucas
+ * @license https://github.com/virginent/yii2-oauth2-server/blob/master/LICENSE
  */
 
-namespace conquer\oauth2\responsetypes;
+namespace virginent\oauth2\responsetypes;
 
-use conquer\oauth2\models\AuthorizationCode;
-use conquer\oauth2\BaseModel;
+use virginent\oauth2\models\AuthorizationCode;
+use virginent\oauth2\BaseModel;
 
 /**
  * @link https://tools.ietf.org/html/rfc6749#section-4.1.1
- * @author Andrey Borodulin
+ * @Author Daniel Lucas
  */
 class Authorization extends BaseModel
 {
@@ -65,7 +65,7 @@ class Authorization extends BaseModel
 
     /**
      * @return array
-     * @throws \conquer\oauth2\Exception
+     * @throws \virginent\oauth2\Exception
      * @throws \yii\base\Exception
      */
     public function getResponseData()