+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 1.7.0
 -----
-- Bug #37: Fixed `conquer\oauth2\TokenAuth` behavior when multiple Authorization headers are present
-- Bug #32: `conquer\oauth2\BaseModel` caches request
+- Bug #37: Fixed `virginent\oauth2\TokenAuth` behavior when multiple Authorization headers are present
+- Bug #32: `virginent\oauth2\BaseModel` caches request
 - Enh #24: Client Credentials grant type support
-- Enh: Added `scopes` setting to `conquer\oauth2\TokenAuth`
+- Enh: Added `scopes` setting to `virginent\oauth2\TokenAuth`
 - Enh: Added migrations namespace